+++ app/server/src/features/accounts/resolvers.ts
@@ -18,6 +18,11 @@ export const resolvers: Resolvers = {
             const user = await User.findUserById(ctx.viewer.id, ctx.prisma);
             return toUserId(user);
         },
+        async isOrganizer(parent, args, ctx, info) {
+            if (!ctx.viewer.id) { return false; }
+            const email = await User.findEmailByUserId(ctx.viewer.id, ctx.prisma);
+            return User.isOnOrganizerList(email?.email!);
+        },
     },
     User: {
         async organizations(parent, args, ctx, info) {
